Fixed link_id problems that were causing extra links to be added
git-svn-id: svn+ssh://wicontroller4/home/svn/wirover_kernel/trunk@220 ba602a36-4df9-46f2-848b-6875ca4683ef

+++ common/ping.h
@@ -10,7 +10,7 @@ struct ping_packet {
     uint8_t     type;
     int8_t      link_state;
     uint16_t    src_id;
-    uint16_t    link_id;
+    uint32_t    link_id;
     int32_t     secret_word;
     uint32_t    sent_sec;
     uint32_t    sent_usec;
